rename: AppTabBarItem 코드 분리(#30)
- AppTabBarItem 코드를 열거형 파일로 분리하고 Enum 그룹으로 이동함
